Create a utils folder for utility functions and a types folder for TypeScript interfaces
